Use UNIX thread_local implementation for WASI.
