+++ library/std/src/sys/thread_local/mod.rs
@@ -85,11 +85,8 @@ pub(crate) mod guard {
         } else if #[cfg(target_os = "windows")] {
             mod windows;
             pub(crate) use windows::enable;
-        } else if #[cfg(all(target_os = "wasi"))] {
-            mod wasi;
-            pub(crate) use wasi::enable;
         } else if #[cfg(any(
-            target_family = "wasm",
+            all(target_family = "wasm", not(target_os="wasi")),
             target_os = "uefi",
             target_os = "zkvm",
         ))] {
@@ -138,6 +135,7 @@ pub(crate) mod key {
                 target_family = "unix",
             ),
             target_os = "teeos",
+            target_os = "wasi",
         ))] {
             mod racy;
             mod unix;
